+++ react/src/components/Banner.js
